makefile works for both
