+++ code/main.cpp
@@ -8,34 +8,7 @@
 // #define DEBUG
 // #define __NVCC__
 
-#ifdef __NVCC__
-#ifdef DEBUG
-#define cudaCheckError(ans) cudaAssert((ans), __FILE__, __LINE__);
-
-inline void cudaAssert(cudaError_t code, const char *file, int line, bool abort=true)
-{
-    if (code != cudaSuccess) {
-        fprintf(stderr, "CUDA Error: %s at %s:%d\n",
-        cudaGetErrorString(code), file, line);
-        if (abort) exit(code);
-    }
-}
-
-#else // __DEBUG__
-
-#define cudaCheckError(ans) ans
-
-#endif // __DEBUG__
-#else // __NVCC__
-#ifdef __JETBRAINS_IDE__
-#include "CLion.h"
-#else
-#define __host__
-#define __device__
-#define __global__
-#endif
-
-#endif // __NVCC__
+#include "Defs.h"
 
 #include "Board2DDevice.h"
 #include "NonogramLineDevice.h"
